requirements.txt
aiohttp
aiosignal
anyio
async-timeout
attrs
beautifulsoup4
beautifultable
bleach
cachetools
certifi
charset-normalizer
#contourpy
cssutils
cycler
dataframe-image
defusedxml
exceptiongroup
fastjsonschema
fonttools
frozenlist
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
gsheet
gspread
h11
html2image
httpcore
httplib2
httpx
idna
importlib-metadata
importlib-resources
Jinja2
jsonschema
jsonschema-specifications
jupyter_client
jupyter_core
jupyterlab-pygments
kaleido
kiwisolver
lxml
MarkupSafe
matplotlib
mistune
multidict
nbclient
nbconvert
nbformat
numpy
oauthlib
packaging
pandas
pandocfilters
Pillow
platformdirs
plotly
protobuf
pyasn1
pyasn1-modules
Pygments
pyparsing
pyTelegramBotAPI
python-dateutil
python-dotenv
python-telegram-bot
pytz
pyzmq
referencing
requests
requests-oauthlib
rpds-py
rsa
six
sniffio
soupsieve
tabulate
tenacity
tinycss2
tornado
traitlets
typing_extensions
tzdata
uritemplate
urllib3
validators
wcwidth
webencodings
websocket-client
yarl
zipp
